Change int types to fit stdint.h naming and improve memory impact

+++ headers/envp.h
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/04 17:21:31 by egualand          #+#    #+#             */
-/*   Updated: 2024/03/01 23:57:55 by craimond         ###   ########.fr       */
+/*   Updated: 2024/03/02 22:41:25 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,6 @@
 # define REMOVE -1
 # define REPLACE 0
 # define ADD 1
-
-// uint8_t hash(char *str);
 
 void  	envp_init(char **envp);
 void	ft_setenv(char *env_name, char *env_value, bool replace);
@@ -31,7 +29,7 @@ t_tree	*envp_tree_find(t_tree *root, char *name, uint16_t name_len);
 void	envp_tree_print(t_tree *root);
 void	envp_tree_print_export(t_tree *root);
 
-void	envp_matrix_remove(char *env_name, uint32_t name_len);
+void	envp_matrix_remove(char *env_name, uint16_t name_len);
 void	envp_matrix_add(char *str);
 void	envp_print_env(void);
 void	envp_print_export(t_tree *root);
